minimega: another round of affinity changes
Seems to fix issues with `optimize affinity`. Moved `VMs.Start` to
`Namespace.Start` so that we can set affinity before starting the VM, if
its enabled.
